Split LSP project analysis support
